Move tests to the sub-project
