--- core/src/test/scala/com/github/gvolpe/fs2rabbit/interpreter/ConnectionStub.scala
@@ -1,32 +1,0 @@
-/*
- * Copyright 2017-2019 Fs2 Rabbit
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.gvolpe.fs2rabbit.interpreter
-
-import cats.effect.IO
-import com.github.gvolpe.fs2rabbit.algebra.Connection
-import com.github.gvolpe.fs2rabbit.model.AMQPChannel
-import com.github.gvolpe.fs2rabbit.effects.StreamEval
-import com.rabbitmq.client.Channel
-import fs2.Stream
-
-class ConnectionStub(implicit SE: StreamEval[IO]) extends Connection[Stream[IO, ?]] {
-
-  case class ChannelStub(value: Channel = null) extends AMQPChannel
-
-  override def createConnectionChannel: Stream[IO, AMQPChannel] = SE.pure(ChannelStub())
-
-}
