+++ test-support/src/main/scala/com/github/gvolpe/fs2rabbit/DockerRabbit.scala
